% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ondo.raw.plot}
\alias{ondo.raw.plot}
\title{温湿度・飽差の時系列データを1日毎にプロットする}
\usage{
ondo.raw.plot(ondo, start = range(ondo$day)[1], end = range(ondo$day)[2])
}
\arguments{
\item{ondo}{\code{\link{read_ocsv_rh}}で読み込んだデータフレーム。}

\item{start}{集計の開始日。\code{"yyyy/mm/dd"}形式で指定する。}

\item{end}{集計の終了日。指定方法はstartと同様。}
}
\description{
\code{\link{read_ocsv_rh}}により読み込んだデータについて、
\code{start}から\code{end}までの期間の温湿度・飽差データのプロットを作成する。
プロットは1日おきに分割され、横に並べられる。
それぞれのプロットには温湿度・飽差の日平均・最高・最低値がテキストとして表示される。
}

