% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ondo.summarized.plot}
\alias{ondo.summarized.plot}
\title{温湿度・飽差の日最高・平均・最低値の推移をプロットする}
\usage{
ondo.summarized.plot(ondo, start = range(ondo$day)[1],
  end = range(ondo$day)[2])
}
\arguments{
\item{ondo}{\code{\link{read_ocsv_rh}}で読み込んだデータフレーム。}

\item{start}{集計の開始日。\code{"yyyy/mm/dd"}形式で指定する。}

\item{end}{集計の終了日。指定方法はstartと同様。}
}
\description{
\code{\link{read_ocsv_rh}}により読み込まれたデータを加工し、
温度、相対湿度、飽差について日最高・平均・最低値を算出し、
これらの推移をプロットする。
}

