% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataio.R
\name{read_ocsv_rh}
\alias{read_ocsv_rh}
\title{「おんどとり」により作成されるデータ(温湿度2ch)を読み込む}
\usage{
read_ocsv_rh(fname)
}
\arguments{
\item{fname}{ファイルネーム。ファイルの形式はカンマ区切りテキスト(CSV)とする。}
}
\value{
返り値はデータフレームであり、以下の列を持つ。
\item{datetime}{POSIXct形式の日時。}
\item{day}{Date形式の日付。}
\item{temp}{気温(C)}
\item{RH}{相対湿度(\%)}
\item{vpd}{飽差(VPD, hPa)}
\item{w.dif}{飽差(HD, g/m^3)}
}
\description{
「おんどとり(T&D)」により採取し、付属ソフトウェアで出力したカンマ区切りテキストファイルを
 読み込み、データフレームに加工する。その際、2種類の飽差(VPD及びHD)を計算し、
 列として付加する。飽差の計算のためには相対湿度情報が必要であるため、
 ch1が温度、ch2が湿度である必要がある。TR-72シリーズなどにより採取されたデータから
 テキスト形式でそのまま出力した場合は該当の形式になっている。
 温度1chのみのデータの読み込みは\code{\link{read_ocsv_s}}を参照。
}

