% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataio.R
\name{read_ocsv_s}
\alias{read_ocsv_s}
\title{「おんどとり」により作成されるデータ(温度1ch)を読み込む}
\usage{
read_ocsv_s(fname)
}
\arguments{
\item{fname}{ファイルネーム。ファイルの形式はカンマ区切りテキスト(CSV)とする。}
}
\value{
返り値はデータフレームであり、以下の列を持つ。
\item{datetime}{POSIXct形式の日時。}
\item{day}{Date形式の日付。}
\item{temp}{気温(C)}
}
\description{
「おんどとり(T&D)」により採取し、付属ソフトウェアで出力したカンマ区切りテキストファイルを
 読み込み、データフレームに加工する。
 ch1に記録されたデータを温度として扱う。
 湿度データをch2に含むデータの読み込みは\code{\link{read_ocsv_rh}}を参照。
}

