% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_value.R
\name{svp}
\alias{svp}
\title{セ氏気温から飽和水蒸気圧を計算する。式はAdduchov and Eskridge(1996)による。}
\usage{
svp(t)
}
\arguments{
\item{t}{気温（セルシウス度）}
}
\value{
飽和水蒸気圧（Pa）
}
\description{
セ氏気温から飽和水蒸気圧を計算する。式はAdduchov and Eskridge(1996)による。
}

