% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_value.R
\name{vpd}
\alias{vpd}
\title{セ氏温度と相対湿度から飽差(VPD)を計算する}
\usage{
vpd(t, RH)
}
\arguments{
\item{t}{気温(セルシウス度)}

\item{RH}{相対湿度(\%)}
}
\value{
飽差(hPa)
}
\description{
セ氏温度と相対湿度から飽差を計算する。
飽差は水蒸気分圧の差(Vapor Pressure Deficit)として求める。
水蒸気量の差として求める飽差(Humidity Deficit)については\code{\link{w.dif}}を参照。
}
\examples{
vpd(25, 50)
vpd(10, 50)
}

