% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_value.R
\name{w.dif}
\alias{w.dif}
\title{セ氏温度と相対湿度から飽差(HD)を計算する}
\usage{
w.dif(t, RH)
}
\arguments{
\item{t}{気温(セルシウス度)}

\item{RH}{相対湿度(\%)}
}
\value{
飽差（g/m^3）
}
\description{
セ氏温度と相対湿度から飽差を計算する。
飽差は水蒸量の差(Humidity Deficit)として求める。
水蒸気分圧の差として求める飽差(Vapor Pressure Deficit)については\code{\link{vpd}}を参照。
}

